flatten handlers folder
